# api/openapi.yaml
openapi: 3.1.0
info:
  title: hostd Phase A API
  version: 0.1.0
  description: Authenticated local-only control plane. Workload execution is disabled by default and requires an explicit runtime capability.
servers:
  - url: http://127.0.0.1:7345
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/apps/{appId}/deployment-plan:
    parameters: [{$ref: '#/components/parameters/AppId'}]
    get:
      operationId: getApplicationDeploymentPlan
      x-rig-problem-codes: [deployment_plan_not_found]
      responses:
        '200': {description: Current accepted deployment plan, headers: {Cache-Control: {schema: {type: string, const: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/DeploymentPlanRevision'}}}}
        default: {$ref: '#/components/responses/Problem'}
    put:
      operationId: acceptApplicationDeploymentPlan
      x-rig-problem-codes: [authentication_required, source_access_lost, provider_unavailable, invalid_source, source_too_large, deployment_plan_conflict, deployment_plan_review_required, invalid_deployment_plan, deployment_plan_forbidden]
      security: [{sessionCookie: [], csrfHeader: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AcceptDeploymentPlanRequest'}}}}
      responses:
        '200': {description: Accepted immutable deployment plan revision, headers: {Cache-Control: {schema: {type: string, const: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/DeploymentPlanRevision'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /api/v1/apps/{appId}/deployment-plan/migration-approval:
    parameters: [{$ref: '#/components/parameters/AppId'}]
    post:
      operationId: approveApplicationDeploymentPlanMigration
      x-rig-problem-codes: [deployment_plan_not_found, deployment_plan_conflict, migration_approval_conflict, invalid_deployment_plan, deployment_plan_forbidden]
      security: [{sessionCookie: [], csrfHeader: []}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ApproveDeploymentPlanMigrationRequest'}}}}
      responses:
        '200': {description: Migration approval recorded, headers: {Cache-Control: {schema: {type: string, const: no-store}}}, content: {application/json: {schema: {$ref: '#/components/schemas/DeploymentPlanRevision'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /api/v1/apps/{appId}/deployments:
    parameters: [{$ref: '#/components/parameters/AppId'}]
    get:
      operationId: listDeployments
      responses:
        '200': {description: Deployment history, content: {application/json: {schema: {$ref: '#/components/schemas/DeploymentList'}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      operationId: deployApplication
      x-rig-problem-codes: [application_busy]
      security: [{sessionCookie: [], csrfHeader: []}]
      parameters: [{$ref: '#/components/parameters/IdempotencyKey'}]
      responses: *jobResponses
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hostd_session
      description: Opaque random session token. Set HttpOnly, Path=/, SameSite=Strict; Secure is enabled when served over TLS. Only a SHA-256 hash is stored.
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Required with sessionCookie for every non-safe request. Rotated by the authenticated CSRF endpoint; only its hash is stored.
...
  parameters:
    AppId: {name: appId, in: path, required: true, schema: {type: string, format: uuid}}
...
    IdempotencyKey: {name: Idempotency-Key, in: header, required: false, schema: {type: string, maxLength: 200}}
...
  responses:
    Problem:
      description: RFC 9457 problem response
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
    Problem:
      type: object
      required: [type, title, status, detail, code, requestId]
      properties:
        type: {type: string, format: uri}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        code: {type: string}
        requestId: {type: string}
        errors: {type: object, additionalProperties: {type: string}}
...
    PolicyFinding: {type: object, required: [id, deploymentId, policyVersion, capability, scope, fingerprint, disposition, createdAt], properties: {id: {type: string, format: uuid}, deploymentId: {type: string, format: uuid}, policyVersion: {type: string}, capability: {type: string}, scope: {type: string}, fingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}, disposition: {type: string, enum: [allowed, approval_required, rejected]}, createdAt: {type: string, format: date-time}}}
    Deployment: {type: object, required: [id, appId, status, configurationMode, actualConfigurationRevisionNumber, runtimeStrategy, deploymentPlanRevisionNumber, findings], properties: {id: {type: string, format: uuid}, appId: {type: string, format: uuid}, releaseId: {type: string}, jobId: {type: string, format: uuid}, machineId: {type: string, format: uuid}, status: {type: string, enum: [preparing, applying, waiting_health, succeeded, failed, cancelled, needs_attention]}, configurationMode: {type: string, enum: [current, original]}, actualConfigurationRevisionId: {type: string, format: uuid}, actualConfigurationRevisionNumber: {type: integer, format: int64, minimum: 0}, runtimeStrategy: {type: string, enum: [compose, generated_node]}, deploymentPlanRevisionId: {type: string, format: uuid}, deploymentPlanRevisionNumber: {type: integer, format: int64, minimum: 0}, startedAt: {type: string, format: date-time}, finishedAt: {type: string, format: date-time}, diagnosticCode: {type: string}, failureSummary: {type: string}, findings: {type: array, items: {$ref: '#/components/schemas/PolicyFinding'}}}}
    DeploymentList: {type: object, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Deployment'}}}}
...
    DeploymentSetupInput:
      type: object
      additionalProperties: false
      required: [components]
      properties:
        components: {type: array, minItems: 1, maxItems: 2, items: {$ref: '#/components/schemas/DeploymentSetupComponentInput'}}
        migrationCommand: {type: string, maxLength: 8192}
    DeploymentSetupComponentInput:
      type: object
      additionalProperties: false
      required: [id, technology, rootDirectory, packageManager, nodeVersion, installCommand, buildCommand, startCommand, outputDirectory, internalPort, healthProbe]
      properties:
        id: {type: string, minLength: 1, maxLength: 256}
        technology: {type: string, enum: [node, nextjs, static]}
        rootDirectory: {type: string, minLength: 1, maxLength: 1024}
        packageManager: {type: string, enum: [npm, pnpm, yarn]}
        nodeVersion: {type: string, enum: ['20', '22', '24']}
        installCommand: {type: string, maxLength: 8192, description: Empty explicitly skips dependency installation.}
        buildCommand: {type: string, maxLength: 8192, description: Empty explicitly skips the build step.}
        startCommand: {type: string, maxLength: 8192, description: Required for servers; empty for managed static sites.}
        outputDirectory: {type: string, maxLength: 1024, description: Required for static sites; relative to rootDirectory.}
        internalPort: {type: integer, minimum: 1, maximum: 65535}
        healthProbe: {type: string, minLength: 1, maxLength: 1024}
    DeploymentPlanComponentInput: {type: object, additionalProperties: false, required: [componentId, buildCommand, runCommand, nodeVersion, internalPort, healthProbe], properties: {componentId: {type: string, minLength: 1, maxLength: 256}, buildCommand: {type: string, maxLength: 8192}, runCommand: {type: string, minLength: 1, maxLength: 8192}, nodeVersion: {type: string, minLength: 1, maxLength: 256}, internalPort: {type: integer, minimum: 1, maximum: 65535}, healthProbe: {type: string, minLength: 1, maxLength: 1024}}}
    AcceptDeploymentPlanRequest: {type: object, additionalProperties: false, required: [expectedRevisionNumber, expectedSourceStructuralFingerprint, expectedCandidateDigest, candidateId], properties: {setup: {$ref: '#/components/schemas/DeploymentSetupInput', x-go-pointer: true}, expectedRevisionNumber: {type: integer, format: int64, minimum: 0}, expectedSourceStructuralFingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}, expectedCandidateDigest: {type: string, pattern: '^[a-f0-9]{64}$'}, candidateId: {type: string, minLength: 1, maxLength: 512}, packageManager: {type: string, enum: [npm, pnpm, yarn]}, installBehavior: {type: string, minLength: 1, maxLength: 8192}, components: {type: array, minItems: 1, maxItems: 64, items: {$ref: '#/components/schemas/DeploymentPlanComponentInput'}}, migrationCommand: {type: string, maxLength: 8192}, migrationEnvironmentKeys: {type: array, maxItems: 8, uniqueItems: true, items: {type: string, pattern: '^[A-Za-z_][A-Za-z0-9_]*$', maxLength: 128}}}}
    ApproveDeploymentPlanMigrationRequest: {type: object, additionalProperties: false, required: [revisionId, revisionNumber, expectedApprovalRevision], properties: {revisionId: {type: string, format: uuid}, revisionNumber: {type: integer, format: int64, minimum: 1}, expectedApprovalRevision: {type: integer, format: int64, minimum: 0}}}
    DeploymentPlanFieldProvenance: {type: object, required: [field, origin, confidence, evidence], properties: {field: {type: string}, origin: {type: string, enum: [inferred, user]}, confidence: {type: integer, minimum: 0, maximum: 100}, evidence: {type: array, items: {type: string}}}}
    DeploymentPlanComponent: {type: object, required: [name, role, rootDirectory, packageManager, installBehavior, installDirectory, nodeVersion, buildCommand, runCommand, internalPort, healthProbe], properties: {name: {type: string}, role: {type: string}, rootDirectory: {type: string}, packageManager: {type: string}, installBehavior: {type: string}, installDirectory: {type: string}, nodeVersion: {type: string}, buildCommand: {type: string}, runCommand: {type: string}, internalPort: {type: integer, minimum: 1, maximum: 65535}, healthProbe: {type: string}}}
    DeploymentPlanDetector: {type: object, required: [name, version, sourceStructuralFingerprint], properties: {name: {type: string}, version: {type: string}, sourceStructuralFingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}}}
    DeploymentPlanMigration: {type: object, required: [present], properties: {present: {type: boolean}, componentName: {type: string}, rootDirectory: {type: string}, command: {type: string}, environmentKeys: {type: array, maxItems: 8, items: {type: string}}, evidenceDigest: {type: string, pattern: '^[a-f0-9]{64}$'}, approvalStatus: {type: string, enum: [pending, approved]}, approvedBy: {type: string, format: uuid}, approvedAt: {type: string, format: date-time}}}
    DeploymentPlanSource: {type: object, required: [provider, repositoryId, resolvedDigest], properties: {provider: {type: string, enum: [local, github]}, repositoryId: {type: integer, format: int64, minimum: 0}, resolvedDigest: {type: string, pattern: '^[a-f0-9]{40}$|^[a-f0-9]{64}$'}}}
    DeploymentPlanRevision: {type: object, required: [revisionNumber, strategy, state, canonicalDigest, source, detector, components, fieldProvenance, migration], properties: {setup: {$ref: '#/components/schemas/DeploymentSetupInput', x-go-pointer: true}, revisionId: {type: string, format: uuid}, revisionNumber: {type: integer, format: int64, minimum: 0}, strategy: {type: string, enum: [generated_node, compose]}, state: {type: string, enum: [accepted, superseded]}, canonicalDigest: {type: string, pattern: '^[a-f0-9]{64}$'}, source: {$ref: '#/components/schemas/DeploymentPlanSource'}, detector: {$ref: '#/components/schemas/DeploymentPlanDetector'}, components: {type: array, items: {$ref: '#/components/schemas/DeploymentPlanComponent'}}, fieldProvenance: {type: array, items: {$ref: '#/components/schemas/DeploymentPlanFieldProvenance'}}, migration: {$ref: '#/components/schemas/DeploymentPlanMigration'}, acceptedBy: {type: string, format: uuid}, acceptedAt: {type: string, format: date-time}}}
...